+++ src/libponyc/codegen/genexe.c
@@ -186,6 +186,19 @@ static void gen_main(compile_t* c, reach_type_t* t_main,
   LLVMSetLinkage(func, LLVMExternalLinkage);
 }
 
+#if defined(PLATFORM_IS_LINUX) || defined(PLATFORM_IS_FREEBSD)
+static const char* env_cc_or_pony_compiler(bool* out_fallback_linker)
+{
+  const char* cc = getenv("CC");
+  if(cc == NULL)
+  {
+    *out_fallback_linker = true;
+    return PONY_COMPILER;
+  }
+  return cc;
+}
+#endif
+
 static bool link_exe(compile_t* c, ast_t* program,
   const char* file_o)
 {
@@ -271,7 +284,16 @@ static bool link_exe(compile_t* c, ast_t* program,
   const char* lib_args = program_lib_args(program);
 
   const char* arch = c->opt->link_arch != NULL ? c->opt->link_arch : PONY_ARCH;
-  const char* linker = c->opt->linker != NULL ? c->opt->linker : PONY_COMPILER;
+  bool fallback_linker = false;
+  const char* linker = c->opt->linker != NULL ? c->opt->linker :
+    env_cc_or_pony_compiler(&fallback_linker);
+
+  if((c->opt->verbosity >= VERBOSITY_MINIMAL) && fallback_linker)
+  {
+    fprintf(stderr,
+      "Warning: environment variable $CC undefined, using %s as the linker\n",
+      PONY_COMPILER);
+  }
   const char* mcx16_arg = target_is_ilp32(c->opt->triple) ? "" : "-mcx16";
   const char* fuseld = target_is_linux(c->opt->triple) ? "-fuse-ld=gold " : "";
   const char* ldl = target_is_linux(c->opt->triple) ? "-ldl  " : "";